chore: rename and restructure project files
