+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
+  "eslint.validate": ["javascript"],
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": "always"
-  },
-  "eslint.validate": ["javascript"]
+  }
 }
