+++ frontend/gameFunctionality/logger.js
