+++ frontend/gamePage.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Cellular conquest</title>
-    <link rel="stylesheet" href="gameStyles.css" />
-    <link rel="stylesheet" href="gameMobile.css" />
+    <link rel="stylesheet" href="./gameStyles/gameStyles.css" />
+    <link rel="stylesheet" href="./gameStyles/gameMobile.css" />
   </head>
   <body>
     <canvas id="gameCanvas"></canvas>
