+++ frontend/gameStyles/gameMobile.css
